// Part 9/patientor-backend/src/types.ts
export interface Diagnose {
  code: string;
  name: string;
  latin?: string;
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

export interface EmptyEntry {
  id?: string;
  description?: string;
  date?: string;
  specialist?: string;
  diagnosisCodes?: string[];
  type?: string;
  healthCheckRating?: HealthCheckRating;
  employerName?: string;
  discharge?: {
    date?: string;
    criteria?: string
  };
}

export interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: {
    startDate?: string;
    endDate?: string;
  };
}

export interface HospitalEntry extends BaseEntry {
  type: "Hospital";
  discharge: {
    date: string;
    criteria: string
  };
}

export type Entry =
  | OccupationalHealthcareEntry
  | HospitalEntry
  | HealthCheckEntry
  | EmptyEntry;

// export type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;
// export type EntryWithoutId = UnionOmit<Entry, 'id'>;

export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn?: string;
  gender: string;
  occupation: string;
  entries?: Entry[];
}

export type NonSensitiveDataPatients = Omit<Patient, 'ssn' | 'entries'>;

export type NewPatientInput = Omit<Patient, 'id'>;

// Define special omit for unions
type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;

// Define Entry without the 'id' property
export type EntryWithoutId = UnionOmit<Entry, 'id'>;

export enum Gender {
  Male = 'male',
  Female = 'female',
  Other = 'other'
}